services/autocut-engine/pyproject.toml
[project]
name = "autocut-engine"
version = "0.1.0"
description = "Auto-Cut media engine"
requires-python = ">=3.12,<3.13"
dependencies = [
  "fastapi==0.116.1",
  "python-dotenv==1.0.1",
  "uvicorn==0.35.0",
  "httpx==0.28.1",
  "pydantic==2.11.7",
  "ffmpeg-python==0.2.0",
  "torch>=2.8.0",
  "whisperx>=3.8.1",
  "hf-xet>=1.3.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
